# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/hei-domains": {
      "get": {
        "operationId": "UsersAdminController_listHeiDomains",
        "summary": "List the institutional email allowlist",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "UsersAdminController_createHeiDomain",
        "summary": "Add an institution to the email allowlist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHeiDomainDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/hei-domains/{id}": {
      "patch": {
        "operationId": "UsersAdminController_setHeiDomainActive",
        "summary": "Enable or disable an allowlisted institution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/pending": {
      "get": {
        "operationId": "UsersAdminController_listPending",
        "summary": "Industry accounts awaiting manual verification",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{id}/activate": {
      "post": {
        "operationId": "UsersAdminController_activate",
        "summary": "Verify and activate a pending account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{id}/suspend": {
      "post": {
        "operationId": "UsersAdminController_suspend",
        "summary": "Suspend an account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/org-units/districts": {
      "get": {
        "operationId": "OrgUnitsController_listDistricts",
        "summary": "List Jharkhand districts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organisation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/org-units/{id}/children": {
      "get": {
        "operationId": "OrgUnitsController_listChildren",
        "summary": "List the direct children of an org unit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organisation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/media/upload-url": {
      "post": {
        "operationId": "MediaController_requestUpload",
        "summary": "Request a presigned upload URL",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media/{id}/confirm": {
      "post": {
        "operationId": "MediaController_confirm",
        "summary": "Confirm the bytes landed and validate the stored object",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media/{id}/download-url": {
      "get": {
        "operationId": "MediaController_download",
        "summary": "Short-lived signed read URL; objects are never public",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "ProjectsController_findAll",
        "summary": "List projects visible to the caller",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "g1_passed",
                "g2_passed",
                "g3_passed",
                "g4_passed",
                "deployed",
                "deployed_verified",
                "on_hold",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "Get a project by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/projects/{id}/issues": {
      "get": {
        "operationId": "ProjectsController_linkedIssues",
        "summary": "Issues linked to this project, including merged ones",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "JEEVAN API",
    "description": "Jharkhand Ecosystem for Experiential Validation Adoption Network",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateHeiDomainDto": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "example": "bitmesra.ac.in"
          },
          "institutionName": {
            "type": "string",
            "example": "Birla Institute of Technology, Mesra"
          },
          "orgUnitId": {
            "type": "string",
            "description": "HEI-tier org unit students are attached to"
          }
        },
        "required": [
          "domain",
          "institutionName"
        ]
      },
      "SetActiveDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "isActive"
        ]
      },
...
      "RequestUploadDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "photo",
              "video",
              "audio",
              "document"
            ]
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "sizeBytes": {
            "type": "number",
            "example": 482913,
            "description": "Declared size; re-checked on confirm"
          },
          "latitude": {
            "type": "number",
            "description": "Capture latitude. Required for photo and video."
          },
          "longitude": {
            "type": "number",
            "description": "Capture longitude. Required for photo and video."
          },
          "capturedAt": {
            "type": "string",
            "example": "2026-08-31T09:15:00Z"
          }
        },
        "required": [
          "kind",
          "mimeType",
          "sizeBytes"
        ]
      },
...
    }
  }
}